1. memcpy

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -1,17 +1,20 @@
 #include "main.h"
 
 /**
- * _memset - Entry point
- * @s: pointed destination
- * @b: constant byte
- * @n: bytes
- * Return: Always 0 (Success)
+ * _memset - this function is semailar to the function memset
+ *
+ * @s: A pointer to the start of the memory block you want to fill.
+ * @b: The value you want to set in each byte of the memory block.
+ * @n: The number of bytes to be set.
+ *
+ * Return: it return the new value of s
  */
+
 char *_memset(char *s, char b, unsigned int n)
 {
-	unsigned int i;
-
+	size_t i;
 	for (i = 0; i < n; i++)
-		s[i] = b;
+	s[i] = b;
+
 	return (s);
 }
